Move PEFA plots into its own component file for clarity
